refactor(models): ProbeStep/SmokeVerdict live in the shared decode loop, gemma4 re-exports
